Move toCent and centToString to one place
