+++ packages/voucherify/data/order-to-voucherify-order.ts
@@ -1,6 +1,6 @@
 import { Order } from '@composable/types'
 import { OrdersCreate } from '@voucherify/sdk'
-import { toCent } from '../src/to-cent'
+import { toCent } from '../../commerce-generic/src/services/cart/to-cent'
 
 export const orderToVoucherifyOrder = (order: Order): OrdersCreate => {
   return {
